requirements.txt
LaTeXCompiler==1.0
Pillow==10.1.0
PySocks==1.7.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
config==0.5.1
distro==1.8.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
openai==1.2.3
outcome==1.3.0.post0
packaging==23.2
pandoc==2.3
pdf2image==1.16.3
pip==23.3.1
plumbum==1.8.2
ply==3.11
pydantic==2.4.2
pydantic-core==2.10.1
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
setuptools==65.5.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing-extensions==4.8.0
urllib3==2.1.0
webdriver-manager==4.0.1
wheel==0.38.4
wsproto==1.2.0